% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_indicator.R
\name{compute_impact_indicator}
\alias{compute_impact_indicator}
\title{Overall impact indicator}
\usage{
compute_impact_indicator(
  cube,
  impact_data = NULL,
  method = NULL,
  trans = 1,
  col_category = NULL,
  col_species = NULL,
  col_mechanism = NULL,
  region = NULL,
  ci_type = c("perc", "bca", "norm", "basic", "none"),
  num_bootstrap = 1000,
  grouping_var = "year",
  no_bias = TRUE,
  out_var = "taxonKey",
  conf = 0.95,
  seed = NA
)
}
\arguments{
\item{cube}{A data cube object (class 'processed_cube' or 'sim_cube', processed
from \code{b3gbi::process_cube()}) or a dataframe (cf. \verb{$data} slot of
'processed_cube' or 'sim_cube') or an impact cube (class 'impact_cube' from
\code{create_impact_cube_data})}

\item{impact_data}{A dataframe of species impact which contains columns of
\code{impact_category}, \code{scientific_name} and \code{impact_mechanism}.}

\item{method}{A method of computing the indicator. The method used in
the aggregation of within and across species in a site proposed
by Boulesnane-Genguant et al. (submitted).
The method can be one of
\itemize{
\item \code{"precaut"}: The "\code{precautionary}" method assigns a species the
maximum impact across all records of the species and then compute the
maximum impact across species in each site
\item \code{"precaut_cum"}: The "\verb{precautionary cumulative}" method assigns a
species the maximum impact across all records of the species then
compute the summation of all impacts in each site.
The precautionary cumulative method provides the highest
impact score possible for each species but considers
the number of co-occurring species in each site.
\item \code{"mean"}:The "\code{mean}" method assigns species the mean impact of all
the species impact and then computes the mean of all species in each site.
The mean provides the expected impact within individual species and
across all species in each site.
\item \code{"mean_cum"}: The "\verb{mean cumulative}" assigns a species the mean
impact of all the species impact and then computes the summation of all
impact scores in each site. The mean cumulative provides the expected
impact score within individual species but adds co-occurring species’
impact scores in each site.
\item \code{"cum"}: The "\code{cumulative}" assigns a species the summation of the
maximum impact per mechanism and then computes the summation of
all species’ impacts per site. The cumulative method provides a
comprehensive view of the overall impact while considering the impact
and mechanisms of multiple species.
}}

\item{trans}{Numeric: \code{1} (default), \code{2} or \code{3}. The method of transformation
to convert the EICAT categories \code{c("MC", "MN", "MO", "MR", "MV")} to
numerical values:
\itemize{
\item \code{1}: converts the categories to \code{c(0, 1, 2, 3, 4)}
\item \code{2}: converts the categories to to \code{c(1, 2, 3, 4, 5)}
\item \code{3}: converts the categories to to \code{c(1, 10, 100, 1000, 10000)}
}}

\item{col_category}{The name of the column containing the impact categories.
The first two letters each categories must be an EICAT short names
(e.g "MC - Minimal concern").}

\item{col_species}{The name of the column containing species names.}

\item{col_mechanism}{The name of the column containing mechanisms of impact.}

\item{region}{The shape file of the specific region to calculate the indicator on.
If NULL (default), the indicator is calculated for all cells in the cube.}

\item{ci_type}{A character vector specifying the type of confidence intervals
to compute. Options include:
\itemize{
\item \code{"perc"}: Percentile interval (Default)
\item \code{"bca"}: Bias-corrected and accelerated interval
\item \code{"norm"}: Normal interval
\item \code{"basic"}: Basic interval
\item \code{"none"}: No intervals are calculated
}}

\item{num_bootstrap}{The number of bootstrap replicates. A single positive integer.
Default is 1000.}

\item{grouping_var}{A character vector specifying the grouping variable(s)
for the bootstrap analysis. The function \code{fun()} should
return a row per group. The specified variables must not be redundant,
meaning they should not contain the same information (e.g., \code{"time_point"}
(1, 2, 3) and \code{"year"} (2000, 2001, 2002) should not be used together if
\code{"time_point"} is just an alternative encoding of \code{"year"}).
This variable is used to split the dataset into groups for separate
confidence interval calculations.}

\item{no_bias}{If TRUE intervals are centered around the original
estimates (bias is ignored). Default is TRUE}

\item{out_var}{A string specifying the column by which the data should be
left out iteratively. Default is \code{"taxonKey"} which can be used for
leave-one-species-out CV.}

\item{conf}{A numeric value specifying the confidence level of the intervals.
Default is 0.95 (95 \% confidence level).}

\item{seed}{A positive numeric value setting the seed for random number
generation to ensure reproducibility. If \code{NA} (default), then \code{set.seed()}
is not called at all. If not \code{NA}, then the random number generator state is
reset (to the state before calling this function) upon exiting this function.}
}
\value{
A list of class \code{impact_indicator}, with the following components:
\itemize{
\item \code{method}: method used in computing the indicator
\item \code{num_cells}: number of cells (sites) in the indicator
\item \code{num_species}:  number of species in the indicator
\item \code{names_species}: names of species in the indicator
\item \code{site_impact}: a dataframe containing total species impact per year
}
}
\description{
Combines occurrences cube and impact data using the given method
(e.g., mean cumulative) to compute the impact indicator of all
species over a given region
}
\examples{
acacia_cube <- taxa_cube(
  taxa = taxa_Acacia,
  region = southAfrica_sf,
  res = 0.25,
  first_year = 2010
)
impact_value <- compute_impact_indicator(
  cube = acacia_cube,
  impact_data = eicat_acacia,
  method = "mean_cum",
  trans = 1,
  ci_type = "none"
)
}
\seealso{
Other Indicator function: 
\code{\link{compute_impact_per_site}()},
\code{\link{compute_impact_per_species}()}
}
\concept{Indicator function}
