% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_impact_cube_data.R
\name{create_impact_cube_data}
\alias{create_impact_cube_data}
\title{Create impact cube data}
\usage{
create_impact_cube_data(
  cube_data,
  impact_data,
  trans = 1,
  col_category = NULL,
  col_species = NULL,
  col_mechanism = NULL,
  region = NULL
)
}
\arguments{
\item{cube_data}{A dataframe of a cube ( $data slot) of class \code{sim_cube} or \code{processed_cube} from
\code{b3gbi::process_cube()}.}

\item{impact_data}{The dataframe of species impact which contains columns of
\code{impact_category}, \code{scientific_name} and \code{impact_mechanism}.}

\item{trans}{Numeric: \code{1} (default), \code{2} or \code{3}. The method of transformation
to convert the EICAT categories \code{c("MC", "MN", "MO", "MR", "MV")} to
numerical values:
\itemize{
\item \code{1}: converts the categories to \code{c(0, 1, 2, 3, 4)}
\item \code{2}: converts the categories to to \code{c(1, 2, 3, 4, 5)}
\item \code{3}: converts the categories to to \code{c(1, 10, 100, 1000, 10000)}
}}

\item{col_category}{The name of the column containing the impact categories.
The first two letters each categories must be an EICAT short names
(e.g "MC - Minimal concern").}

\item{col_species}{The name of the column containing species names.}

\item{col_mechanism}{The name of the column containing mechanisms of impact.}

\item{region}{The shape file of the specific region to calculate the indicator on.
If NULL (default), the indicator is calculated for all cells in the cube.}
}
\value{
A dataframe of containing impact data and occurrence cube. The class
is \code{impact_cube}
}
\description{
This function combines the occurrence cube data and impact data,
named impact_cube_data.
}
\examples{
library(b3gbi) # for processing cube
acacia_cube <- process_cube(cube_name = cube_acacia_SA,
                          grid_type = "eqdgc",
                          first_year = 2010,
                           last_year = 2024)

impact_value <- create_impact_cube_data(
  cube_data = acacia_cube$data,
  impact_data = eicat_acacia,
)

}
