% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cube_acacia_SA}
\alias{cube_acacia_SA}
\title{GBIF Occurrence Cube of acacia in South Africa
An example of occurrence cube of GBIF containing required column for impact
indicator.}
\format{
A dataframe object containing 4,700 rows and 8 variables
\describe{
\item{year}{The year the occurrence was recorded}
\item{eqdcellcode}{The extended quarter degree cell code}
\item{speciesKey}{The GBIF species identification number}
\item{species}{The scientific name of species}
\item{occurrences}{The number of observation in the cell}
\item{kingdom}{The kingdom name of which the species belong}
\item{family}{The family name of which the species belong}
\item{mincoordinateuncertaintyinmeters}{ minimum radius of the uncertainty
circle around the geographic point}
}
}
\source{
GBIF.org (16 October 2025) GBIF Occurrence Download \doi{10.15468/dl.zm3keb}
}
\usage{
cube_acacia_SA
}
\description{
GBIF Occurrence Cube of acacia in South Africa
An example of occurrence cube of GBIF containing required column for impact
indicator.
}
\examples{
head(cube_acacia_SA, 10)

}
\seealso{
Other Data: 
\code{\link{eicat_acacia}},
\code{\link{southAfrica_sf}},
\code{\link{taxa_Acacia}}
}
\concept{Data}
\keyword{datasets}
