% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eicat_acacia}
\alias{eicat_acacia}
\title{EICAT data of acacia taxa
An example of EICAT data containing species name, impact category and mechanism.}
\format{
A dataframe object containing 138 observations and 3 variables
\describe{
\item{scientific_name}{species scientific name}
\item{impact_category}{EICAT impact category}
\item{impact_mechanism}{mecahnism of impact}
}
}
\source{
\url{https://github.com/Oceane-Boulesnane-Guengant/Risk_maps_Acacia_SA/tree/main/data}
}
\usage{
eicat_acacia
}
\description{
EICAT data of acacia taxa
An example of EICAT data containing species name, impact category and mechanism.
}
\examples{
  head(eicat_acacia,10)

}
\keyword{datasets}
