% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eicat_data}
\alias{eicat_data}
\title{GBIF occurrences data of acacia in South Africa
An example of occurrence data from GBIF containing required column for impact indicator.}
\format{
A dataframe object containing 138 observations and 4 variables
\describe{
\item{scientific_name}{species scientific name}
\item{impact_category}{EICAT impact category}
\item{impact_mechanism}{mecahnism of impact}
\item{probability}{weight of impact}
}
}
\source{
\url{https://github.com/Oceane-Boulesnane-Guengant/Risk_maps_Acacia_SA/tree/main/data}
}
\usage{
eicat_data
}
\description{
GBIF occurrences data of acacia in South Africa
An example of occurrence data from GBIF containing required column for impact indicator.
}
\examples{
  head(eicat_data,10)

}
\keyword{datasets}
