% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_category.R
\name{impact_cat}
\alias{impact_cat}
\title{Compute impact categories}
\usage{
impact_cat(
  impact_data,
  species_list,
  col_category = NULL,
  col_species = NULL,
  col_mechanism = NULL,
  trans = 1
)
}
\arguments{
\item{impact_data}{The dataframe of species impact which contains columns of category,
species and mechanism.}

\item{species_list}{The vector of species' list to aggregate their impact categories}

\item{col_category}{The name of the column containing the impact categories.
The first two letters each categories must be an EICAT short names
(e.g "MC -Minimal concern")}

\item{col_species}{The name of the column containing species names}

\item{col_mechanism}{The name of the column containing mechanisms of impact}

\item{trans}{Numeric. The type of transformation to convert the EICAT categories to
numerical values. 1 converts ("MC", "MN", "MO", "MR", "MV") to (0,1,2,3,4)
2 converts ("MC", "MN", "MO", "MR", "MV") to (1,2,3,4,5) and
3 converts ("MC", "MN", "MO", "MR", "MV") to (1,10,100,1000,10000)}
}
\value{
The dataframe containing the aggregated species impact. max - maximum
impact of a species. mean - mean impact of a species. max_mech - sum of maximum
impact per categories of a species
}
\description{
Aggregate species impact categories from impact data.
}
\examples{
#read the impact data
eicat_data<-readRDS("Data/eicat_data.rds")
#define species list
species_list<-c("Acacia adunca",
"Acacia baileyana",
"Acacia binervata",
"Acacia crassiuscula",
"Acacia cultriformis",
"Acacia cyclops",
"Acacia dealbata",
"Acacia decurrens",
"Acacia elata")

agg_impact<-impact_cat(impact_data=eicat_data,
                     species_list=species_list,
                     col_category="impact_category",
                     col_species="scientific_name",
                     col_mechanism="impact_mechanism",
                     trans=1)

}
