% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.site_impact}
\alias{plot.site_impact}
\title{Plot site impact}
\usage{
\method{plot}{site_impact}(
  x,
  region = NULL,
  first_year = NULL,
  last_year = NULL,
  title_lab = "Impact map",
  text_size = 14,
  ...
)
}
\arguments{
\item{x}{A dataframe of impact indicator. Must be a class of "site_impact"}

\item{region}{sf or character. The shapefile of the region of study or a
character which represent the name of a country. It is not compulsory
but makes the plot more comprehensible.}

\item{first_year}{The first year the impact map should include.
Default starts from the first year included in \code{x}.}

\item{last_year}{The last year the impact map should include.
Default ends in the last year included in \code{x}.}

\item{title_lab}{Title of the plot. Default is "Impact map"}

\item{text_size}{The size of the text of the plot. Default is "14"}

\item{...}{Additional arguments passed to geom_tile}
}
\value{
The ggplot of species yearly impact on the region.
}
\description{
Produces the yearly impact map of a region
}
\examples{
# define cube for taxa
acacia_cube <- taxa_cube(
  taxa = taxa_Acacia,
  region = southAfrica_sf,
  res = 0.25,
  first_year = 2010
)

# compute site impact
siteImpact <- site_impact(
  cube = acacia_cube,
  impact_data = eicat_acacia,
  method = "precautionary cumulative",
  trans = 1

)

# visualise site impact
plot(x=siteImpact,
region= southAfrica_sf,
first_year = 2021)
}
