% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.species_impact}
\alias{plot.species_impact}
\title{Plot species impact}
\usage{
\method{plot}{species_impact}(
  x,
  alien_species = "all",
  linewidth = 1.5,
  title_lab = "Species impact indicator",
  y_lab = "Impact score",
  x_lab = "Year",
  text_size = 14,
  ...
)
}
\arguments{
\item{x}{A dataframe of impact indicator. Must be a class of "species_impact"}

\item{alien_species}{The character vector containing names of the alien
species to be included in the plot. Default is "all" which plot all species
in the data frame}

\item{linewidth}{The width size of the line. Default is 1.5}

\item{title_lab}{Title of the plot. Default is "Species impact"}

\item{y_lab}{Label of the y-axis. Default is "impact score"}

\item{x_lab}{Label of the x-axis. Default is "Year"}

\item{text_size}{The size of the text of the plot. Default is "14"}

\item{...}{Additional arguments passed to geom_line}
}
\value{
The ggplot object of the species impact, with the y- and x-axes
representing the impact score and time respectively.
}
\description{
Produces a ggplot to show the trend of the species impact indicator.
}
\examples{
# create data cube
acacia_cube <- taxa_cube(
  taxa = taxa_Acacia,
  region = southAfrica_sf,
  res = 0.25,
  first_year = 2010
)

# compute species impact
speciesImpact <- compute_impact_per_species(
  cube = acacia_cube,
  impact_data = eicat_acacia,
  method = "mean",
  trans = 1
)

# visualise species impact
plot(speciesImpact)
}
\seealso{
Other Plot: 
\code{\link{plot.impact_indicator}()},
\code{\link{plot.site_impact}()}
}
\concept{Plot}
