% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_method.R
\name{print.impact_indicator}
\alias{print.impact_indicator}
\title{Print an impact_indicator object}
\usage{
\method{print}{impact_indicator}(x, n = 10, ...)
}
\arguments{
\item{x}{An impact_indicator object.}

\item{n}{(Optional) Integer specifying the number of rows of data to display.}

\item{...}{Additional arguments.}
}
\description{
Provides a summary representation of an impact_indicator object,
designed for user-friendly display in the console.
}
\examples{
\dontrun{
# create data_cube
acacia_cube <- taxa_cube(
  taxa = taxa_Acacia,
  region = southAfrica_sf,
  res = 0.25,
  first_year = 2010
)

# compute impact indicator
impact_value <- compute_impact_indicator(
  cube = acacia_cube,
  impact_data = eicat_acacia,
  method = "mean_cum",
  trans = 1
)

# print impact indicator
print(impact_value)
}

}
