% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_method.R
\name{print.site_impact}
\alias{print.site_impact}
\title{Print an site_impact object}
\usage{
\method{print}{site_impact}(x, n = 10, ...)
}
\arguments{
\item{x}{An site_impact object.}

\item{n}{(Optional) Integer specifying the number of rows of data to display.}

\item{...}{Additional arguments.}
}
\description{
Provides a summary representation of an site_impact object,
designed for user-friendly display in the console.
}
\examples{
# create data_cube
acacia_cube <- taxa_cube(
  taxa = taxa_Acacia,
  region = southAfrica_sf,
  res = 0.25,
  first_year = 2010
)

siteImpact <- compute_impact_per_site(
  cube = acacia_cube,
  impact_data = eicat_acacia,
  method = "mean_cum",
  trans = 1
)

# print species impact
print(siteImpact)

}
