% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_method.R
\name{print.species_impact}
\alias{print.species_impact}
\title{Print an species_impact object}
\usage{
\method{print}{species_impact}(x, n = 10, ...)
}
\arguments{
\item{x}{An species_impact object.}

\item{n}{(Optional) Integer specifying the number of rows of data to display.}

\item{...}{Additional arguments.}
}
\description{
Provides a summary representation of an species_impact object,
designed for user-friendly display in the console.
}
\examples{
# create data_cube
acacia_cube <- taxa_cube(
  taxa = taxa_Acacia,
  region = southAfrica_sf,
  res = 0.25,
  first_year = 2010
)

# compute species impact
speciesImpact <- compute_impact_per_species(
  cube = acacia_cube,
  impact_data = eicat_acacia,
  method = "mean",
  trans = 1
)

# print species impact
print(speciesImpact)

}
