% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{southAfrica_sf}
\alias{southAfrica_sf}
\title{South African sf
An example of region sf for impact indicator.}
\format{
A 'sf' object of South African map
\describe{
\item{geometry}{geometry of polygon}
}
}
\usage{
southAfrica_sf
}
\description{
South African sf
An example of region sf for impact indicator.
}
\examples{
sf::plot_sf(southAfrica_sf)
}
\seealso{
Other Data: 
\code{\link{cube_acacia_SA}},
\code{\link{eicat_acacia}},
\code{\link{taxa_Acacia}}
}
\concept{Data}
\keyword{datasets}
