% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{species_by_site}
\alias{species_by_site}
\title{Species by site}
\usage{
species_by_site(cube, y)
}
\arguments{
\item{cube}{The list containing data cube of class \code{sim_cube} from
\code{b3gbi::process_cube()}.}

\item{y}{The year to compute it species by site}
}
\value{
A data frame of specie by site. The row represent the site and column
represent the species
}
\description{
Compute species by site from an occurrence cube
}
