% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa_Acacia}
\alias{taxa_Acacia}
\title{GBIF occurrences data of acacia in South Africa
An example of occurrence data from GBIF containing required column for impact indicator.}
\format{
A dataframe object containing 19,100 rows and 6 variables
\describe{
\item{decimalLatitude}{geographic latitude in decimal}
\item{decimalLongitude}{geographic longitude in decimal}
\item{species}{scientific name of species}
\item{speciesKey}{GBIF species identification number}
\item{coordinateUncertaintyInMeters}{radius of the uncertainty circle around geographic point}
\item{year}{year occurrence was recorded}
}
}
\source{
\url{https://doi.org/10.15468/dl.b6gda5}
}
\usage{
taxa_Acacia
}
\description{
GBIF occurrences data of acacia in South Africa
An example of occurrence data from GBIF containing required column for impact indicator.
}
\examples{
  head(taxa_Acacia,10)

}
\keyword{datasets}
