% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_cube.R
\name{taxa_cube}
\alias{taxa_cube}
\title{Prepare Data Cubes}
\usage{
taxa_cube(
  taxa,
  region,
  limit = 500,
  country = NULL,
  res = 0.25,
  first_year = NULL
)
}
\arguments{
\item{taxa}{Character or dataframe. The character should be the scientific
name of the focal taxa while the dataframe is the GBIF occurrences data which must
contain "decimalLatitude","decimalLongitude","species","speciesKey",
"coordinateUncertaintyInMeters","dateIdentified", and "year".}

\item{region}{The shapefile of the region of study}

\item{limit}{Number of records to return from GBIF download.
Default is set to 500}

\item{country}{Two-letter country code (ISO-3166-1) of Country for which
the GBIP occurrences data should be downloaded.}

\item{res}{The resolution of grid cells to be used. Default is 0.25}

\item{first_year}{The year from which the occurrence should start from}
}
\value{
A list containing the \code{sim_cubes} of taxa and the dataframe of
coordinates of sites
}
\description{
Prepare data cube to calculate species impact .
The function \code{taxa_cube} can take in the scientific name of the
taxa of interest as in character or GBIF occurrences data containing
necessary columns. The GBIF occurrences is downloaded if scientific
names is given.
}
\examples{
# read region shapefile
#southAfrica_sf<-readRDS("data/southAfrica_sf.rds")

# load the GBIF occurrence data for taxa
#taxa_Acacia<-readRDS("data/taxa_Acacia.rds")

acacia_cube<-taxa_cube(taxa=taxa_Acacia,
                       region=southAfrica_sf,
                       res=0.25,
                       first_year=2010)

}
