% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_category.R
\name{impact_cat}
\alias{impact_cat}
\title{Compute impact categories}
\usage{
impact_cat(
  impact_data,
  species_list,
  trans = 1,
  col_category = NULL,
  col_species = NULL,
  col_mechanism = NULL
)
}
\arguments{
\item{impact_data}{The dataframe of species impact which contains columns of
\code{impact_category}, \code{scientific_name} and \code{impact_mechanism}.}

\item{species_list}{The vector of species' list to aggregate their impact
categories.}

\item{trans}{Numeric. The method of transformation to convert the EICAT
categories to numerical values. 1 converts ("MC", "MN", "MO", "MR", "MV") to
(0, 1, 2, 3, 4) 2 converts ("MC", "MN", "MO", "MR", "MV") to (1, 2, 3, 4, 5)
and 3 converts ("MC", "MN", "MO", "MR", "MV") to (1, 10, 100, 1000, 10000).}

\item{col_category}{The name of the column containing the impact categories.
The first two letters each categories must be an EICAT short names
(e.g "MC - Minimal concern").}

\item{col_species}{The name of the column containing species names.}

\item{col_mechanism}{The name of the column containing mechanisms of impact.}
}
\value{
The dataframe containing the aggregated species impact. max - maximum
impact of a species. mean - mean impact of a species.
max_mech - sum of maximum impact per categories of a species
}
\description{
Aggregate species impact categories from impact data.
}
\examples{
# define species list
species_list <- c(
  "Acacia adunca",
  "Acacia baileyana",
  "Acacia binervata",
  "Acacia crassiuscula",
  "Acacia cultriformis",
  "Acacia cyclops",
  "Acacia dealbata",
  "Acacia decurrens",
  "Acacia elata"
)

agg_impact <- impact_cat(
  impact_data = eicat_acacia,
  species_list = species_list,
  trans = 1
)

}
\seealso{
Other Prepare data: 
\code{\link{taxa_cube}()}
}
\concept{Prepare data}
