% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.impact_indicator}
\alias{plot.impact_indicator}
\title{Plot impact indicator}
\usage{
\method{plot}{impact_indicator}(
  x,
  linewidth = 2,
  colour = "red",
  title_lab = "Overall impact indicator",
  y_lab = "Impact value",
  x_lab = "Year",
  text_size = 14,
  ...
)
}
\arguments{
\item{x}{A dataframe of impact indicator. Must be a class of
"impact_indicator"}

\item{linewidth}{The width size of the line. Default is 2}

\item{colour}{The colour of the line Default is "red"}

\item{title_lab}{Title of the plot. Default is "Impact indicator"}

\item{y_lab}{Label of the y-axis. Default is "Impact score"}

\item{x_lab}{Label of the x-axis. Default is "Year"}

\item{text_size}{The size of the text of the plot. Default is "14"}

\item{...}{Additional arguments passed to geom_line}
}
\value{
The ggplot object of the impact indicator, with the y- and x-axes
representing the impact score and time respectively.
}
\description{
Produces a ggplot object to show the trend of the impact.
}
\examples{
# create data_cube
acacia_cube <- taxa_cube(
  taxa = taxa_Acacia,
  region = southAfrica_sf,
  res = 0.25,
  first_year = 2010
)

# compute impact indicator
impact_value <- impact_indicator(
  cube = acacia_cube,
  impact_data = eicat_acacia,
  method = "mean_cum",
  trans = 1
)

# plot impact indicator
plot(impact_value)
}
\seealso{
Other Plot: 
\code{\link{plot.site_impact}()},
\code{\link{plot.species_impact}()}
}
\concept{Plot}
