% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_indicators_bootstrap.R
\name{prepare_indicators_bootstrap}
\alias{prepare_indicators_bootstrap}
\title{Prepare indicators for bootstrap}
\usage{
prepare_indicators_bootstrap(
  impact_cube_data,
  indicator = c("overall", "site", "species"),
  indicator_method,
  num_bootstrap = 1000,
  grouping_var = "year",
  ci_type = "perc",
  no_bias = TRUE,
  out_var = "taxonKey",
  conf = 0.95
)
}
\arguments{
\item{impact_cube_data}{An impact cube object (class 'impact_cube' from \code{create_impact_cube_data})}

\item{indicator}{An impact indicator to be computed. Options are 'overall',
'site', and 'species'.}

\item{indicator_method}{A method to compute impact indicator (see methods in \code{impact_indicator})}

\item{num_bootstrap}{The number of bootstrap replicates. A single positive integer.
Default is 1000.}

\item{grouping_var}{A character vector specifying the grouping variable(s)
for the bootstrap analysis. The function \code{fun()} should
return a row per group. The specified variables must not be redundant,
meaning they should not contain the same information (e.g., \code{"time_point"}
(1, 2, 3) and \code{"year"} (2000, 2001, 2002) should not be used together if
\code{"time_point"} is just an alternative encoding of \code{"year"}).
This variable is used to split the dataset into groups for separate
confidence interval calculations.}

\item{ci_type}{A character vector specifying the type of confidence intervals
to compute. Options include:
\itemize{
\item \code{"perc"}: Percentile interval (Default)
\item \code{"bca"}: Bias-corrected and accelerated interval
\item \code{"norm"}: Normal interval
\item \code{"basic"}: Basic interval
}}

\item{no_bias}{Logical. If TRUE intervals are centered around the original estimates (bias is ignored). Default is TRUE}

\item{out_var}{A string specifying the column by which the data should be
left out iteratively. Default is \code{"taxonKey"} which can be used for
leave-one-species-out CV.}

\item{conf}{A numeric value specifying the confidence level of the intervals. Default is 0.95 (95 \% confidence level).}
}
\value{
A named list with three items:
\describe{
\item{bootstrap_params}{List of parameters for \code{bootstrap_cube()}}
\item{ci_params}{List of parameters for \code{calculate_bootstrap_ci()}}
\item{cv_params}{List of parameters for \code{cross_validate_cube()}}
}
}
\description{
Prepare indicators for bootstrap
}
\examples{
library(b3gbi) # for processing cube
acacia_cube <- process_cube(
  cube_name = cube_acacia_SA,
  grid_type = "eqdgc",
  first_year = 2010,
  last_year = 2024
)

impact_cube <- create_impact_cube_data(
  cube_data = acacia_cube$data,
  impact_data = eicat_acacia,
)

prepare_indicators_bootstrap(
  impact_cube_data = impact_cube,
  indicator = "overall",
  indicator_method = "mean_cum"
)
}
