% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_impact.R
\name{site_impact}
\alias{site_impact}
\title{Compute site impact indicator}
\usage{
site_impact(
  cube,
  impact_data = NULL,
  method = NULL,
  trans = 1,
  col_category = NULL,
  col_species = NULL,
  col_mechanism = NULL
)
}
\arguments{
\item{cube}{The data cube of class \code{sim_cube} or \code{processed_cube} from
\code{b3gbi::process_cube()}.}

\item{impact_data}{The dataframe of species impact which contains columns of
\code{impact_category}, \code{scientific_name} and \code{impact_mechanism}.}

\item{method}{The method of computing the indicator. The method used in
the aggregation of within and across species in a site.
The method can be precautionary, precautionary cumulative, mean,
mean cumulative or cumulative.}

\item{trans}{Numeric. The method of transformation to convert the EICAT
categories to numerical values. 1 converts ("MC", "MN", "MO", "MR", "MV") to
(0, 1, 2, 3, 4) 2 converts ("MC", "MN", "MO", "MR", "MV") to (1, 2, 3, 4, 5)
and 3 converts ("MC", "MN", "MO", "MR", "MV") to (1, 10, 100, 1000, 10000).}

\item{col_category}{The name of the column containing the impact categories.
The first two letters each categories must be an EICAT short names
(e.g "MC - Minimal concern").}

\item{col_species}{The name of the column containing species names.}

\item{col_mechanism}{The name of the column containing mechanisms of impact.}
}
\value{
The dataframe of impact indicator per sites (class \code{site_impact})
}
\description{
Compute site impact indicator
}
\examples{
# define cube for taxa
acacia_cube <- taxa_cube(
  taxa = taxa_Acacia,
  region = southAfrica_sf,
  res = 0.25,
  first_year = 2010
)

siteImpact <- site_impact(
  cube = acacia_cube,
  impact_data = eicat_acacia,
  method = "precautionary cumulative",
  trans = 1
)
}
\seealso{
Other Indicator function: 
\code{\link{impact_indicator}()},
\code{\link{species_impact}()}
}
\concept{Indicator function}
