% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_impact.R
\name{species_impact}
\alias{species_impact}
\title{Compute species impact indicator}
\usage{
species_impact(
  cube,
  impact_data = NULL,
  col_category = NULL,
  col_species = NULL,
  col_mechanism = NULL,
  trans = 1,
  type = NULL
)
}
\arguments{
\item{cube}{The data cube of class \code{sim_cube} or
\code{processed_cube} from \code{b3gbi::process_cube()}}

\item{impact_data}{The dataframe of species impact which contains columns of category,
species and mechanism.}

\item{col_category}{The name of the column containing the impact categories.
The first two letters each categories must be an EICAT short names
(e.g "MC - Minimal concern")}

\item{col_species}{The name of the column containing species names}

\item{col_mechanism}{The name of the column containing mechanisms of impact}

\item{trans}{Numeric. The type of transformation to convert the EICAT categories to
numerical values. 1 converts ("MC", "MN", "MO", "MR", "MV") to (0,1,2,3,4)
2 converts ("MC", "MN", "MO", "MR", "MV") to (1,2,3,4,5) and
3 converts ("MC", "MN", "MO", "MR", "MV") to (1,10,100,1000,10000)}

\item{type}{The type indicators based on the aggregation of within and
across species in a site. The type can be precautionary, precautionary cumulative,
mean, mean cumulative or cumulative.}
}
\value{
A dataframe of impact indicator per species (class \code{species_impact})
}
\description{
Compute species impact indicator
}
\examples{

acacia_cube <- taxa_cube(
  taxa = taxa_Acacia,
  region = southAfrica_sf,
  res = 0.25,
  first_year = 2010
)

speciesImpact <- species_impact(
  cube = acacia_cube,
  impact_data = eicat_data,
  col_category = "impact_category",
  col_species = "scientific_name",
  col_mechanism = "impact_mechanism",
  trans = 1,
  type = "mean"
)
}
